Bugfixes and refactoring

+++ app/controllers/businesses_controller.rb
@@ -1,33 +1,8 @@
 class BusinessesController < ApplicationController
+  # Do not wrap params received from post in an additional named hash
+  wrap_parameters false
   before_action :set_business, except: %i[index create search]
   before_action :authenticate, except: %i[index search show]
-
-  def okay
-    [
-      {
-        address: {
-          only: :street,
-          include: [
-            suburb: {
-              only: :name,
-            },
-            postcode: {
-              only: :code,
-            },
-            state: {
-              only: :name,
-            },
-          ],
-        },
-        category: {
-          only: :name,
-        },
-        reviews: {
-          only: :rating,
-        },
-      },
-    ]
-  end
 
   rescue_from ActiveRecord::RecordNotFound do |e|
     render json: { errors: e }, status: :not_found
@@ -163,6 +138,7 @@ class BusinessesController < ApplicationController
   end
 
   def update
+    return unless authorize(@business.user)
     if @business.update(business_params)
       render json: @business
     else
@@ -171,7 +147,7 @@ class BusinessesController < ApplicationController
   end
 
   def destroy
-    # TODO: Replace with a soft-delete, and a restore option?
+    return unless authorize(@business.user)
     if @business.destroy
       render status: :no_content
     else
